refactor(rsc): extract export-all expansion transform
