+++ packages/plugin-rsc/src/transforms/index.ts
@@ -3,3 +3,4 @@ export * from './wrap-export'
 export * from './proxy-export'
 export * from './utils'
 export * from './server-action'
+export * from './expand-export-all'
